requirements.txt
flask
flask-cors
pyjwt
pillow
imgspy
requests
psycopg2